AppData/auto_sanitize/requirements.txt
configparser
RPi.Gpio